+++ lib/src/internal/layout/layout_modifier.dart
@@ -1,6 +1,8 @@
 import 'package:docking/src/layout/docking_layout.dart';
+import 'package:meta/meta.dart';
 
 /// Represents a layout modifier.
+@internal
 abstract class LayoutModifier {
   /// Builds a new root given a layout.
   DockingArea? newLayout(DockingLayout layout);