+++ lib/src/internal/widgets/docking_item_widget.dart
@@ -1,15 +1,17 @@
 import 'package:docking/src/docking_drag.dart';
 import 'package:docking/src/docking_buttons_builder.dart';
 import 'package:docking/src/docking_icons.dart';
+import 'package:docking/src/internal/widgets/draggable_widget.dart';
+import 'package:docking/src/internal/widgets/drop_widget.dart';
 import 'package:docking/src/layout/docking_layout.dart';
 import 'package:docking/src/on_item_close.dart';
 import 'package:docking/src/on_item_selection.dart';
-import 'package:docking/src/widgets/draggable_widget.dart';
-import 'package:docking/src/widgets/drop_widget.dart';
 import 'package:flutter/material.dart';
+import 'package:meta/meta.dart';
 import 'package:tabbed_view/tabbed_view.dart';
 
 /// Represents a widget for [DockingItem].
+@internal
 class DockingItemWidget extends DraggableWidget {
   DockingItemWidget(
       {Key? key,